refactor(turns): dedup run_profile prompt-cluster cross-file copies
- `sanitize_ref_suffix` was byte-for-byte identical in `instruction_bundle.rs`
  and `skill_context.rs`. Move it to `snippet_ref.rs` (`pub(crate)`) — the module
  that already owns the shared ref-hash helpers — and import it in both.
- `instruction_bundle::stable_ref_hash` re-implemented the FNV-1a between-fields
  layout that `snippet_ref::stable_skill_snippet_display_hash` already owns.
  Delegate to it (bit-identical: the pinned ref-hash tests pass unchanged).
- `compute_snapshot_version` / `compute_legacy_snapshot_version` in
  `skill_context.rs` were identical but for one `activation_state` field feed.
  Fold into `compute_snapshot_version_inner(entries, include_activation_state)`;
  the digest byte stream — and every historical snapshot version — is preserved.

ironclaw_turns suite + clippy --all-targets --all-features green.
